requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.0.0
openai>=1.0.0
pygame>=2.5.0
gTTS>=2.3.0 